pyproject.toml
[tool.poetry]
name = "gptelegram"
version = "0.1.0"
description = ""
authors = ["melo-tarcisio <melotarcisio@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
openai = "^0.27.7"
tiktoken = "^0.4.0"
speechrecognition = "^3.10.0"
pydub = "^0.25.1"
pydantic = "^1.10.8"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
